requirements.txt
sqlglot>=26.0.0
mysql-connector-python>=8.0.0
